chore: move TestNavHost to navigation test fixtures
